requirements.txt
google-cloud-bigquery==0.25.0